Refactored React Chakra scaffolding and applied the Factory and Adapter pattern for multi-UI page generation
