+++ lib/pages/adapters/antd.js
@@ -1,4 +1,4 @@
-const { createPages } = require("../page-factory");
+const { createPages } = require("../factory/page-factory");
 const { homeSource } = require("../antd/home");
 const { aboutSource } = require("../antd/about");
 const { dashboardSource } = require("../antd/dashboard");
