+++ lib/pages/adapters/bootstrap.js
@@ -1,4 +1,4 @@
-const { createPages } = require("../page-factory");
+const { createPages } = require("../factory/page-factory");
 const { homeSource } = require("../bootstrap/home");
 const { aboutSource } = require("../bootstrap/about");
 const { dashboardSource } = require("../bootstrap/dashboard");
