+++ lib/pages/adapters/mui.js
@@ -1,4 +1,4 @@
-const { createPages } = require("../page-factory");
+const { createPages } = require("../factory/page-factory");
 const { homeSource } = require("../mui/home");
 const { aboutSource } = require("../mui/about");
 const { dashboardSource } = require("../mui/dashboard");
