+++ lib/pages/adapters/tailwind.js
@@ -1,4 +1,4 @@
-const { createPages } = require("../page-factory");
+const { createPages } = require("../factory/page-factory");
 const { homeSource } = require("../tailwind/home");
 const { aboutSource } = require("../tailwind/about");
 const { dashboardSource } = require("../tailwind/dashboard");
