+++ lib/pages/factory/page-factory.js
@@ -1,6 +1,6 @@
 const fs = require("fs");
 const path = require("path");
-const { writeIfMissing } = require("./utils/page-writer");
+const { writeIfMissing } = require("../../utils/page-writer");
 
 /**
  * Generic page generator (framework-specific sources injected by adapter)
