+++ lib/pages/index.js
@@ -3,4 +3,5 @@ module.exports = {
   ...require("./adapters/bootstrap"),
   ...require("./adapters/antd"),
   ...require("./adapters/tailwind"),
+  ...require("./adapters/chakra"),
 };
