+++ lib/scaffold-react-project-structure.js
@@ -1,7 +1,7 @@
 // lib/scaffold-react-project-structure.js
 const fs = require("fs");
 const path = require("path");
-const { createMUIPages, createBootstrapPages, createTailwindPages, createAntdPages } = require("./pages/index");
+const { createMUIPages, createBootstrapPages, createTailwindPages, createAntdPages, createChakraPages } = require("./pages/index");
 
 function writeFileSafe(p, content) {
   fs.mkdirSync(path.dirname(p), { recursive: true });
@@ -139,663 +139,6 @@ export const router = createBrowserRouter([
     // Default to MUI if no specific UI is chosen
     createMUIPages(pagesDir, extx, ext, useTS);
   }
-}
-
-function createChakraPages(pagesDir, extx, ext, useTS) {
-   const home = path.join(pagesDir, `Home.${extx}`);
-  if (!fs.existsSync(home)) {
-    writeFileSafe(home, `import { Box, Text, SimpleGrid, Container, Card, Heading, Avatar, Button, Stack } from '@chakra-ui/react';
-import { FiTrendingUp, FiUsers, FiZap, FiShield, FiCode, FiFeather } from 'react-icons/fi';
-
-
-export default function Home() {
-  const features = [
-    { icon: <FiTrendingUp />, title: 'Performance', description: 'Lightning-fast builds with modern tooling' },
-    { icon: <FiUsers />, title: 'Team Ready', description: 'Built for collaboration and scalability' },
-    { icon: <FiZap />, title: 'Fast Development', description: 'Hot reload and instant feedback' },
-    { icon: <FiShield />, title: 'Type Safe', description: 'Full TypeScript support with strict mode' },
-    { icon: <FiCode />, title: 'Clean Code', description: 'ESLint + Prettier for code quality' },
-    { icon: <FiFeather />, title: 'Beautiful UI', description: 'Chakra UI components with theming' },
-  ];
-
-  return (
-    <Container maxW="6xl" py={6}>
-      <Box textAlign="center" mb={8}>
-        <Heading as="h1" size="2xl" color="blue.500" mb={2}>
-          Welcome to PolyFront
-        </Heading>
-        <Text fontSize="xl" color="gray.500">
-          Your comprehensive multi-frontend scaffold with everything you need
-        </Text>
-        <Text maxW="3xl" mx="auto" mt={3} color="gray.500">
-          Built with React, TypeScript, Vite, Chakra UI, and modern development practices.
-          Start building your next great application today.
-        </Text>
-      </Box>
-
-      {/* spacing -> gap in v3 */}
-      <SimpleGrid columns={{ base: 1, sm: 2, md: 3 }} gap={5} mb={10}>
-        {features.map((feature, i) => (
-          <Card.Root
-            key={i}
-            transition="all 0.2s"
-            _hover={{ transform: 'translateY(-4px)', boxShadow: 'lg' }}
-          >
-            <Card.Body textAlign="center">
-              {/* v3 Avatar parts */}
-              <Avatar.Root size="xl" mb={3} bg="blue.500" color="white">
-                <Avatar.Fallback>
-                  <Box fontSize="28px" lineHeight="1">
-                    {feature.icon}
-                  </Box>
-                </Avatar.Fallback>
-              </Avatar.Root>
-
-              <Heading as="h3" size="md" mb={1}>
-                {feature.title}
-              </Heading>
-              <Text color="gray.500">{feature.description}</Text>
-            </Card.Body>
-          </Card.Root>
-        ))}
-      </SimpleGrid>
-
-      <Card.Root>
-        <Card.Body textAlign="center">
-          <Heading size="lg" mb={2}>
-            Ready to Get Started?
-          </Heading>
-          <Text color="gray.500" mb={4}>
-            This scaffold includes routing, state management, API integration, testing, and much more.
-            Explore the different pages to see all the features in action.
-          </Text>
-
-          {/* spacing -> gap in v3 */}
-          <Stack direction={{ base: 'column', sm: 'row' }} gap={3} justify="center">
-            {/* In v3 prefer colorPalette; colorScheme also works if you’ve enabled it */}
-            <Button colorPalette="blue" size="lg">
-              View Documentation
-            </Button>
-            <Button variant="outline" size="lg">
-              Learn More
-            </Button>
-          </Stack>
-        </Card.Body>
-      </Card.Root>
-    </Container>
-  );
-}
-`);
-  }
-  // Create other Antd pages
-  createOtherChakraPages(pagesDir, extx, ext, useTS);
-}
-
-function createOtherChakraPages(pagesDir, extx, ext, useTS) {
-   // About page
-  const about = path.join(pagesDir, `About.${extx}`);
-  if (!fs.existsSync(about)) {
-    writeFileSafe(about, `import { Box, Text, Container, Card, Heading, SimpleGrid, Tag, List, Separator, Icon, HStack } from '@chakra-ui/react';
-import { FiCheckCircle, FiStar, FiZap, FiUsers } from 'react-icons/fi';
-
-export default function About() {
-  const techStack = [
-    { name: 'React 19', version: 'Latest', category: 'Framework' },
-    { name: 'TypeScript', version: '5.x', category: 'Language' },
-    { name: 'Vite', version: '7.x', category: 'Build Tool' },
-    { name: 'Chakra UI', version: '3.x', category: 'UI Library' },
-    { name: 'React Router', version: '7.x', category: 'Routing' },
-    { name: 'Zustand', version: '5.x', category: 'State Management' },
-    { name: 'Axios', version: '1.x', category: 'HTTP Client' },
-    { name: 'Vitest', version: '3.x', category: 'Testing' },
-  ];
-
-  const features = [
-    'Modern React with hooks and functional components',
-    'Full TypeScript support with strict mode',
-    'Accessible, themable Chakra UI components',
-    'Client-side routing with React Router',
-    'State management with Zustand',
-    'API integration with Axios',
-    'Comprehensive testing setup with Vitest',
-    'ESLint and Prettier for code quality',
-    'Hot module replacement for fast development',
-    'Responsive design with mobile-first approach',
-  ];
-
-  return (
-    <Container maxW="6xl" py={6}>
-      <Box textAlign="center" mb={8}>
-        <Heading as="h1" size="2xl" color="blue.500" mb={2}>
-          About PolyFront
-        </Heading>
-        <Text fontSize="xl" color="gray.500">
-          A comprehensive scaffold for modern web applications
-        </Text>
-      </Box>
-
-      <SimpleGrid columns={{ base: 1, md: 2 }} gap={5} mb={8}>
-        <Card.Root>
-          <Card.Header>
-            <Heading size="md" color="blue.500">
-              Technology Stack
-            </Heading>
-          </Card.Header>
-          <Card.Body>
-            <SimpleGrid columns={{ base: 2, sm: 3 }} gap={3}>
-              {techStack.map((t, i) => (
-                <Box key={i}>
-                  <Tag.Root colorPalette="blue" variant="subtle" mb={1}>
-                    {t.name} {t.version}
-                  </Tag.Root>
-                  <Text fontSize="xs" color="gray.500">
-                    {t.category}
-                  </Text>
-                </Box>
-              ))}
-            </SimpleGrid>
-          </Card.Body>
-        </Card.Root>
-
-        <Card.Root>
-          <Card.Header>
-            <Heading size="md" color="blue.500">
-              Key Features
-            </Heading>
-          </Card.Header>
-          <Card.Body>
-            <List.Root>
-              {features.map((f, i) => (
-                <List.Item key={i}>
-                  <HStack gap={2} align="start">
-                    <Icon as={FiCheckCircle} color="green.400" mt={1} />
-                    <Text>{f}</Text>
-                  </HStack>
-                </List.Item>
-              ))}
-            </List.Root>
-          </Card.Body>
-        </Card.Root>
-      </SimpleGrid>
-
-      <Card.Root>
-        <Card.Body textAlign="center">
-          <SimpleGrid columns={{ base: 1, sm: 3 }} gap={6}>
-            <Box>
-              <Icon as={FiStar} boxSize={10} color="blue.500" mb={1} />
-              <Heading size="sm" mb={1}>
-                Production Ready
-              </Heading>
-              <Text fontSize="sm" color="gray.500">
-                Built with best practices and industry standards
-              </Text>
-            </Box>
-            <Box>
-              <Icon as={FiZap} boxSize={10} color="blue.500" mb={1} />
-              <Heading size="sm" mb={1}>
-                Fast Development
-              </Heading>
-              <Text fontSize="sm" color="gray.500">
-                Hot reload and instant feedback for rapid iteration
-              </Text>
-            </Box>
-            <Box>
-              <Icon as={FiUsers} boxSize={10} color="blue.500" mb={1} />
-              <Heading size="sm" mb={1}>
-                Team Friendly
-              </Heading>
-              <Text fontSize="sm" color="gray.500">
-                Consistent code style and comprehensive tooling
-              </Text>
-            </Box>
-          </SimpleGrid>
-        </Card.Body>
-      </Card.Root>
-
-      <Separator mt={8} />
-    </Container>
-  );
-}`);
-  }
-
-  // Dashboard page
-  const dashboard = path.join(pagesDir, `Dashboard.${extx}`);
-  if (!fs.existsSync(dashboard)) {
-    writeFileSafe(dashboard, `import {
-      Box,
-      Heading,
-      Text,
-      SimpleGrid,
-      Card,
-      Avatar,
-      HStack,
-      VStack,
-      Badge,
-      Table,
-      Icon,
-    } from '@chakra-ui/react';
-    import { FiTrendingUp, FiUsers, FiShoppingCart, FiEye, FiCheckCircle, FiAlertTriangle, FiXCircle } from 'react-icons/fi';
-    
-    export default function Dashboard() {
-      const stats = { users: 1234, revenue: 45678, orders: 89, views: 12345 };
-    
-      const recentActivity = [
-        { id: 1, user: 'Nirmal Samaranayaka', action: 'Order placed', time: '2 min ago', status: 'success'  ${useTS?' as const':''} },
-        { id: 2, user: 'Jane Smith', action: 'Payment received', time: '5 min ago', status: 'success'  ${useTS?' as const':''} },
-        { id: 3, user: 'Bob Johnson', action: 'Login attempt', time: '10 min ago', status: 'warning'  ${useTS?' as const':''} },
-        { id: 4, user: 'Alice Brown', action: 'Account created', time: '15 min ago', status: 'success'  ${useTS?' as const':''} },
-      ];
-    
-      const topProducts = [
-        { name: 'Premium Widget', sales: 156, revenue: 2340, growth: 12 },
-        { name: 'Super Gadget', sales: 98, revenue: 1890, growth: -5 },
-        { name: 'Mega Tool', sales: 87, revenue: 1450, growth: 8 },
-        { name: 'Ultra Device', sales: 76, revenue: 1230, growth: 15 },
-      ];
-    
-      function StatusIcon({ status }${useTS? `: { status: 'success' | 'warning' | 'error' }`:''}) {
-        if (status === 'success') return <Icon as={FiCheckCircle} color="green.400" />;
-        if (status === 'warning') return <Icon as={FiAlertTriangle} color="yellow.400" />;
-        return <Icon as={FiXCircle} color="red.400" />;
-      }
-    
-      return (
-        <Box>
-          <Heading as="h1" size="2xl" color="blue.500" mb={2}>
-            Dashboard
-          </Heading>
-          <Text color="gray.500" mb={6}>
-            Welcome back! Here's what's happening with your application today.
-          </Text>
-    
-          <SimpleGrid columns={{ base: 1, sm: 2, md: 4 }} gap={4} mb={6}>
-            <Card.Root>
-              <Card.Body>
-                <HStack justify="space-between">
-                  <Box>
-                    <Text fontSize="sm" color="gray.500">Total Users</Text>
-                    <Heading size="lg">{stats.users.toLocaleString()}</Heading>
-                    <HStack color="green.500" fontSize="sm" gap={1}>
-                      <Icon as={FiTrendingUp} />
-                      <Text>+12% from last month</Text>
-                    </HStack>
-                  </Box>
-                  <Avatar.Root>
-                    <Avatar.Fallback>
-                      <Icon as={FiUsers} />
-                    </Avatar.Fallback>
-                  </Avatar.Root>
-                </HStack>
-              </Card.Body>
-            </Card.Root>
-    
-            <Card.Root>
-              <Card.Body>
-                <HStack justify="space-between">
-                  <Box>
-                    <Text fontSize="sm" color="gray.500">Revenue</Text>
-                    <Heading size="lg">{stats.revenue.toLocaleString()}</Heading>
-                    <HStack color="green.500" fontSize="sm" gap={1}>
-                      <Icon as={FiTrendingUp} />
-                      <Text>+8% from last month</Text>
-                    </HStack>
-                  </Box>
-                  <Avatar.Root>
-                    <Avatar.Fallback>💰</Avatar.Fallback>
-                  </Avatar.Root>
-                </HStack>
-              </Card.Body>
-            </Card.Root>
-    
-            <Card.Root>
-              <Card.Body>
-                <HStack justify="space-between">
-                  <Box>
-                    <Text fontSize="sm" color="gray.500">Orders</Text>
-                    <Heading size="lg">{stats.orders}</Heading>
-                    <HStack color="green.500" fontSize="sm" gap={1}>
-                      <Icon as={FiTrendingUp} />
-                      <Text>+15% from last month</Text>
-                    </HStack>
-                  </Box>
-                  <Avatar.Root>
-                    <Avatar.Fallback>
-                      <Icon as={FiShoppingCart} />
-                    </Avatar.Fallback>
-                  </Avatar.Root>
-                </HStack>
-              </Card.Body>
-            </Card.Root>
-    
-            <Card.Root>
-              <Card.Body>
-                <HStack justify="space-between">
-                  <Box>
-                    <Text fontSize="sm" color="gray.500">Page Views</Text>
-                    <Heading size="lg">{stats.views.toLocaleString()}</Heading>
-                    <HStack color="green.500" fontSize="sm" gap={1}>
-                      <Icon as={FiTrendingUp} />
-                      <Text>+22% from last month</Text>
-                    </HStack>
-                  </Box>
-                  <Avatar.Root>
-                    <Avatar.Fallback>
-                      <Icon as={FiEye} />
-                    </Avatar.Fallback>
-                  </Avatar.Root>
-                </HStack>
-              </Card.Body>
-            </Card.Root>
-          </SimpleGrid>
-    
-          <SimpleGrid columns={{ base: 1, md: 3 }} gap={4}>
-            <Card.Root gridColumn={{ base: 'span 1', md: 'span 2' }}>
-              <Card.Header>
-                <Heading size="md">Top Products</Heading>
-              </Card.Header>
-              <Card.Body>
-                <Table.Root size="sm">
-                  <Table.Header>
-                    <Table.Row>
-                      <Table.ColumnHeader>Product</Table.ColumnHeader>
-                      <Table.ColumnHeader textAlign="end">Sales</Table.ColumnHeader>
-                      <Table.ColumnHeader textAlign="end">Revenue</Table.ColumnHeader>
-                      <Table.ColumnHeader textAlign="end">Growth</Table.ColumnHeader>
-                    </Table.Row>
-                  </Table.Header>
-                  <Table.Body>
-                    {topProducts.map((p) => (
-                      <Table.Row key={p.name}>
-                        <Table.Cell>{p.name}</Table.Cell>
-                        <Table.Cell textAlign="end">{p.sales}</Table.Cell>
-                        <Table.Cell textAlign="end">{p.revenue.toLocaleString()}</Table.Cell>
-                        <Table.Cell textAlign="end">
-                          <Badge colorPalette={p.growth >= 0 ? 'green' : 'red'}>
-                            {p.growth >= 0 ? '+' : ''}
-                            {p.growth}%
-                          </Badge>
-                        </Table.Cell>
-                      </Table.Row>
-                    ))}
-                  </Table.Body>
-                </Table.Root>
-              </Card.Body>
-            </Card.Root>
-    
-            <Card.Root>
-              <Card.Header>
-                <Heading size="md">Recent Activity</Heading>
-              </Card.Header>
-              <Card.Body>
-                <VStack align="stretch" gap={3}>
-                  {recentActivity.map((a) => (
-                    <HStack key={a.id} align="start" gap={3}>
-                      <Avatar.Root>
-                        <Avatar.Fallback>
-                          <StatusIcon status={a.status} />
-                        </Avatar.Fallback>
-                      </Avatar.Root>
-                      <Box>
-                        <Text fontWeight="medium">{a.action}</Text>
-                        <Text fontSize="sm" color="gray.500">
-                          {a.user} — {a.time}
-                        </Text>
-                      </Box>
-                    </HStack>
-                  ))}
-                </VStack>
-              </Card.Body>
-            </Card.Root>
-          </SimpleGrid>
-        </Box>
-      );
-    }
-    `);
-  }
-
-  // Profile page
-  const profile = path.join(pagesDir, `Profile.${extx}`);
-    if (!fs.existsSync(profile)) {
-      writeFileSafe(profile, `import { useState } from 'react';
-import {
-  Box,
-  Heading,
-  Text,
-  SimpleGrid,
-  Card,
-  Avatar,
-  Button,
-  Field,
-  Input,
-  Textarea,
-  Tag,
-  HStack,
-  Icon,
-  VStack,
-  Switch,
-} from '@chakra-ui/react';
-import { FiEdit, FiSave, FiX, FiBell, FiShield, FiMoon } from 'react-icons/fi';
-
-export default function Profile() {
-  const [isEditing, setIsEditing] = useState(false);
-  const [settings, setSettings] = useState({
-    emailNotifications: true,
-    pushNotifications: false,
-    darkMode: false,
-    twoFactorAuth: true,
-  });
-
-  const profile = {
-    name: 'Nirmal Samaranayaka',
-    email: 'nirmal.fullstack@gmail.com',
-    phone: '+46 (72) xxx-xxxx',
-    location: 'Stockholm, Sweden',
-    company: 'Scania AB.',
-    position: 'Senior Fullstack Developer',
-    education: 'Computer Science, University of Colombo',
-    website: 'https://dev.to/nirmalsamaranayaka',
-    bio: 'Experienced Full Stack Engineer & Tech Lead | Specialized in .NET, React, Angular, and scalable cloud-native solutions.',
-  };
-
-  const skills = [
-    'React', 'Redux', 'MobX', 'React Query', 'Angular', 'Micro Frontends', 'TypeScript', 'JavaScript',
-    'HTML/CSS', 'MUI', 'Tailwind', '.NET', 'C#', 'Azure', 'Docker', 'Kubernetes',
-  ];
-
-  return (
-    <Box>
-      <HStack justify="space-between" mb={4}>
-        <Heading as="h1" size="2xl" color="blue.500">
-          Profile
-        </Heading>
-
-        {isEditing ? (
-          <HStack gap={3}>
-            <Button colorPalette="blue" onClick={() => setIsEditing(false)}>
-              <HStack gap={2}>
-                <FiSave />
-                <span>Save</span>
-              </HStack>
-            </Button>
-            <Button variant="outline" onClick={() => setIsEditing(false)}>
-              <HStack gap={2}>
-                <FiX />
-                <span>Cancel</span>
-              </HStack>
-            </Button>
-          </HStack>
-        ) : (
-          <Button colorPalette="blue" onClick={() => setIsEditing(true)}>
-            <HStack gap={2}>
-              <FiEdit />
-              <span>Edit Profile</span>
-            </HStack>
-          </Button>
-        )}
-      </HStack>
-
-      <SimpleGrid columns={{ base: 1 }} gap={4}>
-        <Card.Root>
-          <Card.Body>
-            <HStack align="start" gap={6}>
-              <Avatar.Root size="2xl">
-                <Avatar.Fallback>
-                  {profile.name.split(' ').map((n) => n[0]).join('').slice(0, 2)}
-                </Avatar.Fallback>
-              </Avatar.Root>
-
-              <Box flex="1">
-                <Heading size="lg" mb={1}>{profile.name}</Heading>
-                <Text color="gray.600" mb={3}>
-                  {profile.position} at {profile.company}
-                </Text>
-                <Text color="gray.600" mb={3}>{profile.bio}</Text>
-
-                <HStack gap={2} flexWrap="wrap">
-                  {skills.map((s) => (
-                    <Tag.Root key={s} variant="outline" colorPalette="blue">
-                      {s}
-                    </Tag.Root>
-                  ))}
-                </HStack>
-              </Box>
-            </HStack>
-          </Card.Body>
-        </Card.Root>
-
-        <SimpleGrid columns={{ base: 1, md: 3 }} gap={4}>
-          <Card.Root gridColumn={{ base: 'span 1', md: 'span 2' }}>
-            <Card.Header>
-              <Heading size="md">Personal Information</Heading>
-            </Card.Header>
-            <Card.Body>
-              <SimpleGrid columns={{ base: 1, sm: 2 }} gap={4}>
-                <Field.Root>
-                  <Field.Label>Full Name</Field.Label>
-                  <Input value={profile.name} disabled={!isEditing} />
-                </Field.Root>
-
-                <Field.Root>
-                  <Field.Label>Email</Field.Label>
-                  <Input value={profile.email} disabled={!isEditing} />
-                </Field.Root>
-
-                <Field.Root>
-                  <Field.Label>Phone</Field.Label>
-                  <Input value={profile.phone} disabled={!isEditing} />
-                </Field.Root>
-
-                <Field.Root>
-                  <Field.Label>Location</Field.Label>
-                  <Input value={profile.location} disabled={!isEditing} />
-                </Field.Root>
-
-                <Field.Root>
-                  <Field.Label>Company</Field.Label>
-                  <Input value={profile.company} disabled={!isEditing} />
-                </Field.Root>
-
-                <Field.Root>
-                  <Field.Label>Position</Field.Label>
-                  <Input value={profile.position} disabled={!isEditing} />
-                </Field.Root>
-
-                <Field.Root gridColumn={{ base: 'span 1', sm: 'span 2' }}>
-                  <Field.Label>Bio</Field.Label>
-                  <Textarea value={profile.bio} disabled={!isEditing} rows={3} />
-                </Field.Root>
-              </SimpleGrid>
-            </Card.Body>
-          </Card.Root>
-
-          <Card.Root>
-            <Card.Header>
-              <Heading size="md">Settings</Heading>
-            </Card.Header>
-            <Card.Body>
-              <VStack align="stretch" gap={4}>
-                <HStack justify="space-between">
-                  <HStack gap={2}>
-                    <Icon as={FiBell} />
-                    <Text>Email Notifications</Text>
-                  </HStack>
-                  <Switch.Root
-                    checked={settings.emailNotifications}
-                    onCheckedChange={(e) =>
-                      setSettings((s) => ({ ...s, emailNotifications: e.checked }))
-                    }
-                  >
-                    {/* hide the built-in checkbox to avoid double rendering */}
-                    <Switch.HiddenInput style={{ display: 'none' }} />
-                    <Switch.Control>
-                      <Switch.Thumb />
-                    </Switch.Control>
-                  </Switch.Root>
-                </HStack>
-
-                <HStack justify="space-between">
-                  <HStack gap={2}>
-                    <Icon as={FiBell} />
-                    <Text>Push Notifications</Text>
-                  </HStack>
-                  <Switch.Root
-                    checked={settings.pushNotifications}
-                    onCheckedChange={(e) =>
-                      setSettings((s) => ({ ...s, pushNotifications: e.checked }))
-                    }
-                  >
-                    <Switch.HiddenInput style={{ display: 'none' }} />
-                    <Switch.Control>
-                      <Switch.Thumb />
-                    </Switch.Control>
-                  </Switch.Root>
-                </HStack>
-
-                <HStack justify="space-between">
-                  <HStack gap={2}>
-                    <Icon as={FiMoon} />
-                    <Text>Dark Mode</Text>
-                  </HStack>
-                  <Switch.Root
-                    checked={settings.darkMode}
-                    onCheckedChange={(e) =>
-                      setSettings((s) => ({ ...s, darkMode: e.checked }))
-                    }
-                  >
-                    <Switch.HiddenInput style={{ display: 'none' }} />
-                    <Switch.Control>
-                      <Switch.Thumb />
-                    </Switch.Control>
-                  </Switch.Root>
-                </HStack>
-
-                <HStack justify="space-between">
-                  <HStack gap={2}>
-                    <Icon as={FiShield} />
-                    <Text>Two-Factor Auth</Text>
-                  </HStack>
-                  <Switch.Root
-                    checked={settings.twoFactorAuth}
-                    onCheckedChange={(e) =>
-                      setSettings((s) => ({ ...s, twoFactorAuth: e.checked }))
-                    }
-                  >
-                    <Switch.HiddenInput style={{ display: 'none' }} />
-                    <Switch.Control>
-                      <Switch.Thumb />
-                    </Switch.Control>
-                  </Switch.Root>
-                </HStack>
-              </VStack>
-            </Card.Body>
-          </Card.Root>
-        </SimpleGrid>
-      </SimpleGrid>
-    </Box>
-  );
-}`);
-    }
-
 }
 
 function createGlobalCSS(srcDir, ui) {
